# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGR \u2014 Agentic Governance Runtime",
    "description": "Drop-in governance layer for AI agent frameworks. Every agent tool call is evaluated against Cedar policies before execution. Sensitive actions require human approval.\n\n## Authentication\n\nAll endpoints (except `/v1/health`) require `Authorization: Bearer agr_sk_<key>` header.\n\n## Decision Values\n\n- `ALLOW` \u2014 agent may proceed\n- `DENY` \u2014 action is blocked\n- `APPROVAL_REQUIRED` \u2014 action is queued for human approval",
    "contact": {
      "name": "Shreeja AI",
      "url": "https://shreejaai.com/"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/org/sso": {
      "get": {
        "tags": [
          "org"
        ],
        "summary": "Get Org Sso Settings",
        "operationId": "get_org_sso_settings_v1_org_sso_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SSOSettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "org"
        ],
        "summary": "Update Org Sso Settings",
        "operationId": "update_org_sso_settings_v1_org_sso_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SSOSettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SSOSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/compliance/summary": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Compliance Summary",
        "description": "Return an aggregated compliance posture for the last N days.\n\nComputes:\n- Total evaluation count and decision breakdown\n- High-risk event count (payload.risk_level == \"high\")\n- Compliance findings aggregated from stored payload data\n\nCompliance findings are stored in audit event payloads during /evaluate.\nThis endpoint aggregates them across recent events.",
        "operationId": "compliance_summary_v1_compliance_summary_get",
        "parameters": [
          {
            "name": "period_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 7,
              "title": "Period Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/compliance/export": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Export compliance report",
        "description": "Export a compliance posture report for the requested period. **Admin role required.** Supported formats: `json` (default), `csv`, and `pdf` (falls back to JSON when PDF rendering is unavailable). The JSON export includes org info, risk summary, audit trail summary, policy violations, per-standard pass/fail counts, and a row per compliance finding.",
        "operationId": "export_compliance_v1_compliance_export_get",
        "parameters": [
          {
            "name": "period_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 7,
              "title": "Period Days"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv|pdf)$",
              "default": "json",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplianceSummaryResponse": {
        "properties": {
          "period_days": {
            "type": "integer",
            "title": "Period Days"
          },
          "total_evaluations": {
            "type": "integer",
            "title": "Total Evaluations"
          },
          "decisions": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Decisions"
          },
          "high_risk_count": {
            "type": "integer",
            "title": "High Risk Count"
          },
          "findings_by_standard": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "type": "object",
            "title": "Findings By Standard"
          },
          "overall_pass": {
            "type": "boolean",
            "title": "Overall Pass"
          },
          "compliance_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Compliance Score",
            "description": "Aggregate compliance score across sampled findings."
          }
        },
        "type": "object",
        "required": [
          "period_days",
          "total_evaluations",
          "decisions",
          "high_risk_count",
          "findings_by_standard",
          "overall_pass",
          "compliance_score"
        ],
        "title": "ComplianceSummaryResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SSOSettingsResponse": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "metadata_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Url"
          },
          "metadata_xml": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Xml"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Domains"
          },
          "default_role": {
            "type": "string",
            "enum": [
              "admin",
              "operator",
              "viewer"
            ],
            "title": "Default Role",
            "default": "viewer"
          },
          "auto_join": {
            "type": "boolean",
            "title": "Auto Join",
            "default": false
          }
        },
        "type": "object",
        "title": "SSOSettingsResponse"
      },
      "SSOSettingsUpdateRequest": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "metadata_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Url"
          },
          "metadata_xml": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Xml"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "domains": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domains"
          },
          "default_role": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "admin",
                  "operator",
                  "viewer"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Role"
          },
          "auto_join": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Join"
          }
        },
        "type": "object",
        "title": "SSOSettingsUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}